requirements.txt
argon2-cffi==23.1.0
pynacl==1.5.0
fastapi==0.110.0
uvicorn==0.29.0
requests==2.31.0
pydantic==2.6.4